+++ area.sh
@@ -12,8 +12,8 @@ case "$1" in
         type docker-compose >/dev/null 2>&1 || { echo >&2 "docker-compose is required but it's not installed.  Aborting."; exit 1; }
         docker-compose -f docker-compose-test.yml build && docker-compose -f docker-compose-test.yml up --abort-on-container-exit
         ;;
-  *)
-        echo "Usage: controlTower.sh {test-e2e|test-unit|start|develop|test}" >&2
+    *)
+        echo "Usage: area.sh {start|develop|test}" >&2
         exit 1
         ;;
 esac